refactor(#50): move test helpers to testhelpers_test.go
